contract_ui/requirements.txt
Flask
requests
gunicorn
python-docx
reportlab
markdown
python-dotenv
